removed file input, redux works well with drag and drop, added shit to input text and password and added links (not styled yet)
