+++ src/App.css
@@ -68,3 +68,8 @@ body {
   border: 5px dotted black;
   border-radius: 20px;
 }
+
+#test {
+  height: 100%;
+  width: 100%;
+}
