+++ src/components/styled/atoms/inputStyled.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import styled from 'styled-components'
 
 const InputStyled = styled.input`
@@ -11,12 +11,26 @@ ${ props => props.inputStyles }
 `
 
 const InputWrapper = ( {type, inputStyles, text, ...props }) => {
+    const [value, setValue] = useState('')
+
+    useEffect(() => {
+        switch (type) {
+          case 'text':
+          case 'password':
+              console.log(value)
+              break
+          default:
+              break
+        }
+    }, [value, type])
+
     return (
         <InputStyled
           type={ type }
           inputStyles = { inputStyles }
           value = { text }
           placeholder = { props.placeholder }
+          onChange = { e => setValue(e.target.value) }
         />
     )
 }
