+++ src/redux/actions/addFileAction.js
@@ -1,5 +1,4 @@
 export function addOneFileAction(file) {
-    console.log('singleFile')
     return {
         type: 'NEW_FILE',
         payload: file
